create destination.rs
